backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM Integration
huggingface-hub==0.20.3
requests==2.31.0

# Twilio (WhatsApp)
twilio==8.11.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Optional (for Groq & Google)
groq==0.4.0
google-generativeai==0.3.2
